Ranamed ToonEvAdjustmentCurve and its inspector classes.
