+++ HDRP-DefaultVolumeProject-2020.3/Assets/Scenes/ExposureCheckScene.unity
@@ -1734,7 +1734,6 @@ MonoBehaviour:
   m_ToonLightHiCutFilter: 1
   m_ExposureAdjustmnt: 1
   m_IgnorVolumeExposure: 1
-  m_HighCutFilter: 1000000
   m_AnimationCurve:
     serializedVersion: 2
     m_Curve:
@@ -1748,8 +1747,8 @@ MonoBehaviour:
       inWeight: 0
       outWeight: 0
     - serializedVersion: 3
-      time: -3.549773
-      value: -6.9683843
+      time: -3.7667682
+      value: -4.5156255
       inSlope: 0.9999999
       outSlope: 0.9999999
       tangentMode: 0
@@ -1770,133 +1769,133 @@ MonoBehaviour:
     m_RotationOrder: 4
   m_ExposureArray:
   - -10
-  - -9.999702
-  - -9.998809
-  - -9.997315
-  - -9.99522
-  - -9.992519
-  - -9.989211
-  - -9.985292
-  - -9.980762
-  - -9.975614
-  - -9.969847
-  - -9.963459
-  - -9.956448
-  - -9.948809
-  - -9.940539
-  - -9.931638
-  - -9.922102
-  - -9.911927
-  - -9.901112
-  - -9.889652
-  - -9.877547
-  - -9.864793
-  - -9.851386
-  - -9.837326
-  - -9.822607
-  - -9.807229
-  - -9.791187
-  - -9.774481
-  - -9.757105
-  - -9.7390585
-  - -9.720338
-  - -9.70094
-  - -9.680863
-  - -9.660105
-  - -9.63866
-  - -9.6165285
-  - -9.593706
-  - -9.57019
-  - -9.5459795
-  - -9.521069
-  - -9.495457
-  - -9.46914
-  - -9.442117
-  - -9.414383
-  - -9.385937
-  - -9.356775
-  - -9.326896
-  - -9.296295
-  - -9.264971
-  - -9.232921
-  - -9.200141
-  - -9.166629
-  - -9.132382
-  - -9.097399
-  - -9.061674
-  - -9.025207
-  - -8.987994
-  - -8.950032
-  - -8.91132
-  - -8.871853
-  - -8.831629
-  - -8.790646
-  - -8.748899
-  - -8.706388
-  - -8.66311
-  - -8.6190605
-  - -8.574238
-  - -8.528639
-  - -8.482261
-  - -8.4351015
-  - -8.387157
-  - -8.338427
-  - -8.288905
-  - -8.238591
-  - -8.187482
-  - -8.135574
-  - -8.082866
-  - -8.029353
-  - -7.9750338
-  - -7.9199057
-  - -7.8639655
-  - -7.80721
-  - -7.7496376
-  - -7.691244
-  - -7.6320276
-  - -7.5719852
-  - -7.511114
-  - -7.4494123
-  - -7.386875
-  - -7.3235025
-  - -7.2592893
-  - -7.194234
-  - -7.1283326
-  - -7.0615845
-  - -6.9939847
-  - -6.9219875
-  - -6.8327656
-  - -6.7261915
-  - -6.6034465
-  - -6.465713
-  - -6.3141713
-  - -6.150003
-  - -5.974389
-  - -5.788512
-  - -5.593553
-  - -5.3906927
-  - -5.1811132
-  - -4.965995
-  - -4.746521
-  - -4.5238705
-  - -4.2992277
-  - -4.0737715
-  - -3.848684
-  - -3.625147
-  - -3.404342
-  - -3.1874506
-  - -2.975653
-  - -2.770131
-  - -2.5720668
-  - -2.3826418
-  - -2.2030363
-  - -2.034432
-  - -1.8780117
-  - -1.7349548
-  - -1.6064434
-  - -1.49366
-  - -1.3977842
-  - -1.3199992
+  - -9.998777
+  - -9.995135
+  - -9.98911
+  - -9.98074
+  - -9.970062
+  - -9.957114
+  - -9.941935
+  - -9.924559
+  - -9.905025
+  - -9.883371
+  - -9.859634
+  - -9.833853
+  - -9.806062
+  - -9.776301
+  - -9.744608
+  - -9.711019
+  - -9.675571
+  - -9.638304
+  - -9.599252
+  - -9.558455
+  - -9.515949
+  - -9.471774
+  - -9.425964
+  - -9.378559
+  - -9.329596
+  - -9.27911
+  - -9.227141
+  - -9.173727
+  - -9.118904
+  - -9.062709
+  - -9.005181
+  - -8.946356
+  - -8.886273
+  - -8.824968
+  - -8.76248
+  - -8.698845
+  - -8.6341
+  - -8.568285
+  - -8.501435
+  - -8.433589
+  - -8.364783
+  - -8.295056
+  - -8.224444
+  - -8.152987
+  - -8.080719
+  - -8.007679
+  - -7.933905
+  - -7.859434
+  - -7.784304
+  - -7.708552
+  - -7.6322145
+  - -7.5553303
+  - -7.477936
+  - -7.40007
+  - -7.3217688
+  - -7.2430706
+  - -7.164012
+  - -7.0846314
+  - -7.0049667
+  - -6.925053
+  - -6.84493
+  - -6.764634
+  - -6.6842036
+  - -6.603675
+  - -6.5230856
+  - -6.4424744
+  - -6.3618774
+  - -6.281333
+  - -6.200878
+  - -6.12055
+  - -6.040387
+  - -5.9604254
+  - -5.880703
+  - -5.801258
+  - -5.722127
+  - -5.643348
+  - -5.5649586
+  - -5.486995
+  - -5.4094963
+  - -5.3324995
+  - -5.2560415
+  - -5.180159
+  - -5.104892
+  - -5.030276
+  - -4.9563484
+  - -4.8831477
+  - -4.8107104
+  - -4.7390747
+  - -4.668278
+  - -4.598357
+  - -4.5293493
+  - -4.4598794
+  - -4.387151
+  - -4.311269
+  - -4.232429
+  - -4.1508274
+  - -4.06666
+  - -3.9801228
+  - -3.8914118
+  - -3.800722
+  - -3.7082498
+  - -3.6141913
+  - -3.518742
+  - -3.4220982
+  - -3.324455
+  - -3.226009
+  - -3.1269555
+  - -3.0274909
+  - -2.9278107
+  - -2.8281112
+  - -2.7285876
+  - -2.6294358
+  - -2.5308523
+  - -2.4330325
+  - -2.3361726
+  - -2.2404683
+  - -2.1461148
+  - -2.053309
+  - -1.9622462
+  - -1.8731225
+  - -1.7861335
+  - -1.7014754
+  - -1.6193433
+  - -1.5399342
+  - -1.463443
+  - -1.3900664
+  - -1.3199997
   m_Max: -1.32
   m_Min: -10
   m_DebugUI: 0
